courier_service/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
pymysql
cryptography
web3
py-solc-x